pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "icon-skill"
version = "0.1.0"
description = "Liquid Glass App Icon Generator — ultra-minimal flat silhouette icons for Apple Icon Composer"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.0.0",
    "Pillow>=10.0.0",
]

[project.scripts]
icon-skill = "icon_skill.main:main"
